refactor:#39 Movie Type 중복 수정
